+++ cmd/peggo/keys.go
@@ -2,19 +2,13 @@ package main
 
 import (
 	"bytes"
-	"crypto/rand"
 	"fmt"
 	"io"
-	"log"
 	"math/big"
 	"os"
 	"strings"
 	"syscall"
 
-	cosmcrypto "github.com/cosmos/cosmos-sdk/crypto"
-	"github.com/cosmos/cosmos-sdk/crypto/keyring"
-	cryptotypes "github.com/cosmos/cosmos-sdk/crypto/types"
-	cosmtypes "github.com/cosmos/cosmos-sdk/types"
 	"github.com/ethereum/go-ethereum/accounts"
 	"github.com/ethereum/go-ethereum/accounts/abi/bind"
 	"github.com/ethereum/go-ethereum/accounts/usbwallet"
@@ -26,15 +20,8 @@ import (
 	"github.com/pkg/errors"
 	"golang.org/x/crypto/ssh/terminal"
 
-	"github.com/InjectiveLabs/sdk-go/chain/crypto/hd"
-
 	"github.com/InjectiveLabs/peggo/orchestrator/ethereum/keystore"
 )
-
-const defaultKeyringKeyName = "validator"
-
-var emptyCosmosAddress = cosmtypes.AccAddress{}
-var cdc = MakeEncodingConfig().Marshaler
 
 var emptyEthAddress = ethcmn.Address{}
 
@@ -252,33 +239,4 @@ func (r *passReader) Read(p []byte) (n int, err error) {
 	}
 
 	return
-}
-
-// KeyringForPrivKey creates a temporary in-mem keyring for a PrivKey.
-// Allows to init Context when the key has been provided in plaintext and parsed.
-func KeyringForPrivKey(name string, privKey cryptotypes.PrivKey) (keyring.Keyring, error) {
-	kb := keyring.NewInMemory(cdc, hd.EthSecp256k1Option())
-	tmpPhrase := randPhrase(64)
-	armored := cosmcrypto.EncryptArmorPrivKey(privKey, tmpPhrase, privKey.Type())
-	err := kb.ImportPrivKey(name, armored, tmpPhrase)
-	if err != nil {
-		err = errors.Wrap(err, "failed to import privkey")
-		return nil, err
-	}
-
-	return kb, nil
-}
-
-func randPhrase(size int) string {
-	buf := make([]byte, size)
-	_, err := rand.Read(buf)
-	orPanic(err)
-
-	return string(buf)
-}
-
-func orPanic(err error) {
-	if err != nil {
-		log.Panicln()
-	}
 }
