+++ orchestrator/cosmos/network.go
@@ -2,6 +2,7 @@ package cosmos
 
 import (
 	"context"
+	"fmt"
 	"time"
 
 	comethttp "github.com/cometbft/cometbft/rpc/client/http"
@@ -18,6 +19,14 @@ import (
 	"github.com/InjectiveLabs/peggo/orchestrator/cosmos/tendermint"
 	"github.com/InjectiveLabs/peggo/orchestrator/ethereum/keystore"
 )
+
+type NetworkConfig struct {
+	ChainID,
+	ValidatorAddress,
+	CosmosGRPC,
+	TendermintRPC,
+	GasPrice string
+}
 
 type Network interface {
 	peggy.QueryClient
@@ -92,4 +101,46 @@ func awaitConnection(client chain.ChainClient, timeout time.Duration) *grpc.Clie
 			return grpcConn
 		}
 	}
+}
+
+func (cfg NetworkConfig) loadClientConfig() clientcommon.Network {
+	if custom := cfg.CosmosGRPC != "" && cfg.TendermintRPC != ""; custom {
+		return customEndpoints(cfg)
+	}
+
+	return loadBalancedEndpoints(cfg)
+}
+
+func customEndpoints(cfg NetworkConfig) clientcommon.Network {
+	c := clientcommon.LoadNetwork("devnet", "")
+	c.Name = "custom"
+	c.ChainId = cfg.ChainID
+	c.Fee_denom = "inj"
+	c.TmEndpoint = cfg.TendermintRPC
+	c.ChainGrpcEndpoint = cfg.CosmosGRPC
+	c.ExplorerGrpcEndpoint = ""
+	c.LcdEndpoint = ""
+	c.ExplorerGrpcEndpoint = ""
+
+	log.Infoln("using custom endpoints for Injective")
+
+	return c
+}
+
+func loadBalancedEndpoints(cfg NetworkConfig) clientcommon.Network {
+	var networkName string
+	switch cfg.ChainID {
+	case "injective-1":
+		networkName = "mainnet"
+	case "injective-777":
+		networkName = "devnet"
+	case "injective-888":
+		networkName = "testnet"
+	default:
+		panic(fmt.Errorf("no provider for chain id %s", cfg.ChainID))
+	}
+
+	log.Infoln("using load balanced endpoints for Injective")
+
+	return clientcommon.LoadNetwork(networkName, "lb")
 }
